pyproject.toml
[project]
name = "vpppn"
version = "0.1.0"
description = "WireGuard control bot/backend skeleton (aiogram + Postgres)"
authors = [{ name = "vpppn maintainers" }]
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.4.1",
    "SQLAlchemy==2.0.27",
    "asyncpg==0.29.0",
    "pydantic==2.5.3",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
